Work around release_test task not working on Ruby 1.x
